+++ src/component/subComponents/RubyPreview.jsx
