+++ src/component/subComponents/SQLPreview.jsx
